requirements.txt
opencv-python-headless>=4.8.0
fer>=22.5.0
Pillow>=10.0.0
numpy>=1.24.0
mcp>=1.0.0
setuptools<81
